requirements.txt
flask>=3.0.0
openai>=1.30.0
python-dotenv>=1.0.0
markdown>=3.5.0
gunicorn>=21.2.0
pytest>=8.0.0
pytest-cov>=5.0.0
